requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-chroma
langchain-openai
python-dotenv
tiktoken
pydantic
openai
marked
lucide
